chore(msw): move fixture for reuse.
