+++ src/__fixtures__/msw.ts
@@ -1,0 +1,74 @@
+import { http, HttpResponse, type HttpHandler } from "msw";
+import { STUB_CAPS_RESPONSE } from "./caps";
+import { SYSTEM_PROMPTS } from "./prompts";
+import { STUB_LINKS_FOR_CHAT_RESPONSE } from "./chat_links_response";
+import { CHAT_LINKS_URL } from "../services/refact/consts";
+
+export const goodPing: HttpHandler = http.get(
+  "http://127.0.0.1:8001/v1/ping",
+  () => {
+    return HttpResponse.text("pong");
+  },
+);
+
+export const goodCaps: HttpHandler = http.get(
+  "http://127.0.0.1:8001/v1/caps",
+  () => {
+    return HttpResponse.json(STUB_CAPS_RESPONSE);
+  },
+);
+
+export const noTools: HttpHandler = http.get(
+  "http://127.0.0.1:8001/v1/tools",
+  () => {
+    return HttpResponse.json([]);
+  },
+);
+
+export const goodPrompts: HttpHandler = http.get(
+  "http://127.0.0.1:8001/v1/customization",
+  () => {
+    return HttpResponse.json({ system_prompts: SYSTEM_PROMPTS });
+  },
+);
+
+export const noCompletions: HttpHandler = http.post(
+  "http://127.0.0.1:8001/v1/at-command-completion",
+  () => {
+    return HttpResponse.json({
+      completions: [],
+      replace: [0, 0],
+      is_cmd_executable: false,
+    });
+  },
+);
+
+export const noCommandPreview: HttpHandler = http.post(
+  "http://127.0.0.1:8001/v1/at-command-preview",
+  () => {
+    return HttpResponse.json({
+      messages: [],
+    });
+  },
+);
+
+export const goodUser: HttpHandler = http.get(
+  "https://www.smallcloud.ai/v1/login",
+  () => {
+    return HttpResponse.json({
+      retcode: "OK",
+      account: "party@refact.ai",
+      inference_url: "https://www.smallcloud.ai/v1",
+      inference: "PRO",
+      metering_balance: -100000,
+      questionnaire: {},
+    });
+  },
+);
+
+export const chatLinks: HttpHandler = http.post(
+  `http://127.0.0.1:8001${CHAT_LINKS_URL}`,
+  () => {
+    return HttpResponse.json(STUB_LINKS_FOR_CHAT_RESPONSE);
+  },
+);
